core/requirements.txt
requests
kafka-python
praw
langchain-google-genai
langgraph
